+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -1,21 +1,11 @@
 package roomescape.theme.controller;
 
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.global.auth.annotation.Admin;
 import roomescape.global.dto.response.ApiResponse;
-import roomescape.theme.dto.ThemeRequest;
-import roomescape.theme.dto.ThemeResponse;
 import roomescape.theme.dto.ThemesResponse;
 import roomescape.theme.service.ThemeService;
 
@@ -41,29 +31,5 @@ public class ThemeController {
             @NotNull(message = "날짜는 null일 수 없습니다.") final LocalDate today
     ) {
         return ApiResponse.success(themeService.findTop10Themes(today));
-    }
-
-    @Admin
-    @PostMapping("/admin/themes")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ApiResponse<ThemeResponse> saveTheme(
-            @Valid @RequestBody final ThemeRequest request,
-            final HttpServletResponse response
-    ) {
-        ThemeResponse themeResponse = themeService.addTheme(request);
-        response.setHeader(HttpHeaders.LOCATION, "/themes/" + themeResponse.id());
-
-        return ApiResponse.success(themeResponse);
-    }
-
-    @Admin
-    @DeleteMapping("/admin/themes/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public ApiResponse<Void> removeTheme(
-            @NotNull(message = "themeId는 null 일 수 없습니다.") @PathVariable final Long id
-    ) {
-        themeService.removeThemeById(id);
-
-        return ApiResponse.success();
     }
 }
